Extract view model from CardReadersView
